+++ lib/ecc/api.dart
@@ -157,9 +157,24 @@ class ECPublicKey extends ECAsymmetricKey implements PublicKey {
 /// A [Signature] created with ECC.
 class ECSignature implements Signature {
   final BigInt r;
-  final BigInt s;
+  BigInt s;
 
   ECSignature(this.r, this.s);
+
+  /**
+   * 'normalize' this signature by converting its s to lower-s form if necessary
+   * This is required to validate this signature with some libraries such as libsecp256k1
+   * which enforce lower-s form for all signatures to combat ecdsa signature malleability
+   *
+   * Returns false if the signature was already normalized, or true if it changed
+   */
+  bool normalize(ECDomainParameters curveParams) {
+    if (s.compareTo(curveParams.n >> 1) > 0) {
+      s = curveParams.n - s;
+      return true;
+    }
+    return false;
+  }
 
   String toString() => "(${r.toString()},${s.toString()})";